+++ application/views/admin/item/itemEdit.php
@@ -4,20 +4,21 @@
 
     <section class="content-header">
         <h1>
-            <i class="fa fa-object-group"></i> <?php echo $this->lang->line('inventory'); ?></h1>
+            <i class="fa fa-object-group"></i> <?php echo $this->lang->line('inventory'); ?>
+        </h1>
     </section>
 
     <!-- Main content -->
     <section class="content">
         <div class="row">
-            <?php if ($this->rbac->hasPrivilege('item', 'can_add') || $this->rbac->hasPrivilege('item', 'can_edit')) { ?> 
+            <?php if ($this->rbac->hasPrivilege('item', 'can_add') || $this->rbac->hasPrivilege('item', 'can_edit')) { ?>
                 <div class="col-md-4">
                     <!-- Horizontal Form -->
                     <div class="box box-primary">
                         <div class="box-header with-border">
                             <h3 class="box-title"><?php echo $this->lang->line('edit_item'); ?></h3>
                         </div><!-- /.box-header -->
-                        <form id="form1" action="<?php echo site_url('admin/item/edit/' . $id) ?>"  id="employeeform" name="employeeform" method="post" accept-charset="utf-8" >
+                        <form id="form1" action="<?php echo site_url('admin/item/edit/' . $id) ?>" id="employeeform" name="employeeform" method="post" accept-charset="utf-8">
                             <div class="box-body">
 
                                 <?php if ($this->session->flashdata('msg')) { ?>
@@ -30,26 +31,35 @@
                                 ?>
                                 <?php echo $this->customlib->getCSRF(); ?>
                                 <input type="hidden" name="id" value="<?php echo $item['id']; ?>">
+                                <div class='form-group'>
+                                    <label for='exampleInputEmail1'><?php echo $this->lang->line('branch'); ?></label><small class='req'> *</small>
+                                    <select id='branch_id' name='branch_id' placeholder='' type='text' class='form-control'>
+                                        <?php foreach ($branch as $key => $value) {  ?>
+                                            <option value='<?php echo $value['id'] ?>' <?php if($item['branch_id'] == $value['id']) { echo "selected"; } ?> ><?php echo $value['branch_name'] ?></option>
+                                        <?php } ?>
+                                    </select>
+                                    <span class='text-danger'><?php echo form_error('branch'); ?></span>
+                                </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('name'); ?></label><small class="req"> *</small>
-                                    <input autofocus="" id="name" name="name" placeholder="" type="text" class="form-control"  value="<?php echo set_value('name', $item['name']); ?>" />
+                                    <input autofocus="" id="name" name="name" placeholder="" type="text" class="form-control" value="<?php echo set_value('name', $item['name']); ?>" />
                                     <span class="text-danger"><?php echo form_error('name'); ?></span>
                                 </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('item_category'); ?></label><small class="req"> *</small>
 
-                                    <select  id="item_category_id" name="item_category_id" class="form-control" >
+                                    <select id="item_category_id" name="item_category_id" class="form-control">
                                         <option value=""><?php echo $this->lang->line('select'); ?></option>
                                         <?php
                                         foreach ($itemcatlist as $item_category) {
-                                            ?>
-                                            <option value="<?php echo $item_category['id'] ?>"<?php
-                                            if (set_value('item_category_id', $item['item_category_id']) == $item_category['id']) {
-                                                echo "selected = selected";
-                                            }
-                                            ?>><?php echo $item_category['item_category'] ?></option>
+                                        ?>
+                                            <option value="<?php echo $item_category['id'] ?>" <?php
+                                                                                                if (set_value('item_category_id', $item['item_category_id']) == $item_category['id']) {
+                                                                                                    echo "selected = selected";
+                                                                                                }
+                                                                                                ?>><?php echo $item_category['item_category'] ?></option>
 
-                                            <?php
+                                        <?php
                                             $count++;
                                         }
                                         ?>
@@ -58,7 +68,7 @@
                                 </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('unit'); ?></label><small class="req"> *</small>
-                                    <input autofocus="" id="unit" name="unit" placeholder="" type="text" class="form-control"  value="<?php echo set_value('unit', $item['unit']); ?>" />
+                                    <input autofocus="" id="unit" name="unit" placeholder="" type="text" class="form-control" value="<?php echo set_value('unit', $item['unit']); ?>" />
                                     <span class="text-danger"><?php echo form_error('unit'); ?></span>
                                 </div>
                                 <div class="form-group">
@@ -74,16 +84,17 @@
                         </form>
                     </div>
 
-                </div><!--/.col (right) -->
+                </div>
+                <!--/.col (right) -->
                 <!-- left column -->
             <?php } ?>
             <div class="col-md-<?php
-            if ($this->rbac->hasPrivilege('item', 'can_add') || $this->rbac->hasPrivilege('item', 'can_edit')) {
-                echo "8";
-            } else {
-                echo "12";
-            }
-            ?> ">
+                                if ($this->rbac->hasPrivilege('item', 'can_add') || $this->rbac->hasPrivilege('item', 'can_edit')) {
+                                    echo "8";
+                                } else {
+                                    echo "12";
+                                }
+                                ?> ">
                 <!-- general form elements -->
                 <div class="box box-primary">
                     <div class="box-header ptbnull">
@@ -97,7 +108,7 @@
                             <table class="table table-hover table-striped table-bordered example">
                                 <thead>
                                     <tr>
-                                        <th><?php echo $this->lang->line('item'); ?></th>                                 
+                                        <th><?php echo $this->lang->line('item'); ?></th>
                                         <th><?php echo $this->lang->line('category'); ?>
                                         </th>
                                         <th><?php echo $this->lang->line('unit'); ?>
@@ -110,12 +121,12 @@
                                 <tbody>
                                     <?php
                                     if (empty($itemlist)) {
-                                        ?>
+                                    ?>
 
                                         <?php
                                     } else {
                                         foreach ($itemlist as $items) {
-                                            ?>
+                                        ?>
                                             <tr>
                                                 <td class="mailbox-name">
                                                     <a href="#" data-toggle="popover" class="detail_popover"><?php echo $items['name'] ?></a>
@@ -123,13 +134,13 @@
                                                     <div class="fee_detail_popover" style="display: none">
                                                         <?php
                                                         if ($items['description'] == "") {
-                                                            ?>
+                                                        ?>
                                                             <p class="text text-danger"><?php echo $this->lang->line('no_description'); ?></p>
-                                                            <?php
+                                                        <?php
                                                         } else {
-                                                            ?>
+                                                        ?>
                                                             <p class="text text-info"><?php echo $items['description']; ?></p>
-                                                            <?php
+                                                        <?php
                                                         }
                                                         ?>
                                                     </div>
@@ -146,8 +157,7 @@
                                                 </td>
                                                 <td class="mailbox-name">
                                                     <?php
-                                                    echo $items['added_stock'] - $items['issued'];
-                                                    ;
+                                                    echo $items['added_stock'] - $items['issued'];;
                                                     ?>
 
                                                 </td>
@@ -156,19 +166,20 @@
 
 
                                                 <td class="mailbox-date pull-right">
-                                                    <?php if ($this->rbac->hasPrivilege('item', 'can_edit')) { ?> 
-                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/item/edit/<?php echo $items['id'] ?>" class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
+                                                    <?php if ($this->rbac->hasPrivilege('item', 'can_edit')) { ?>
+                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/item/edit/<?php echo $items['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
                                                             <i class="fa fa-pencil"></i>
                                                         </a>
-                                                    <?php }if ($this->rbac->hasPrivilege('item', 'can_delete')) { ?>  
-                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/item/delete/<?php echo $items['id'] ?>" class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
+                                                    <?php }
+                                                    if ($this->rbac->hasPrivilege('item', 'can_delete')) { ?>
+                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/item/delete/<?php echo $items['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
                                                             <i class="fa fa-remove"></i>
                                                         </a>
                                                     <?php } ?>
 
                                                 </td>
                                             </tr>
-                                            <?php
+                                    <?php
                                         }
                                     }
                                     ?>
@@ -181,7 +192,8 @@
                         </div><!-- /.mail-box-messages -->
                     </div><!-- /.box-body -->
                 </div>
-            </div><!--/.col (left) -->
+            </div>
+            <!--/.col (left) -->
             <!-- right column -->
 
         </div>
@@ -191,18 +203,19 @@
             <!-- right column -->
             <div class="col-md-12">
 
-            </div><!--/.col (right) -->
-        </div>   <!-- /.row -->
+            </div>
+            <!--/.col (right) -->
+        </div> <!-- /.row -->
     </section><!-- /.content -->
 </div><!-- /.content-wrapper -->
 <script>
-    $(document).ready(function () {
+    $(document).ready(function() {
         $('.detail_popover').popover({
             placement: 'right',
             trigger: 'hover',
             container: 'body',
             html: true,
-            content: function () {
+            content: function() {
                 return $(this).closest('td').find('.fee_detail_popover').html();
             }
         });
